Updated Motocross project

+++ app/Http/Controllers/BookingController.php
@@ -19,7 +19,10 @@ class BookingController extends Controller
      */
     public function index(): View|Application|Factory|\Illuminate\Contracts\Foundation\Application
     {
-        return view('bookings.index', ['bookings' => Booking::with('user', 'schedule')->get()]);
+        $bookings = Auth::user()->bookings;
+
+        if (Auth::user()->role === 'admin') return view('bookings.index', ['bookings' => Booking::all()]);
+        else return view('bookings.index', compact('bookings'));
     }
 
     /**
@@ -36,18 +39,41 @@ class BookingController extends Controller
      * @param Request $request
      * @return RedirectResponse
      */
-    public function store(Request $request, Booking $booking): RedirectResponse
+    public function store(Request $request): RedirectResponse
     {
         $request->validate([
             'schedule_id' => 'required|exists:schedules,id'
         ]);
+
+        // Vérifie si l'utilisateur a déjà une réservation pour ce créneau horaire
+        if (Booking::where('user_id', Auth::user()->id)->where('schedule_id', $request->schedule_id)->exists()) {
+            return redirect()->route('bookings.index')
+                ->with('warning', 'Vous êtes déjà inscrit à cet entraînement.');
+        }
+
+        $schedule = Schedule::find($request->schedule_id);
+        $maxPeople = $schedule->training->max_people;
+
+        // Compte le nombre de réservations pour le créneau horaire actuel
+        $currentBookingsCount = Booking::where('schedule_id', $request->schedule_id)->count() + 1;
+        $remainingPlaces = $maxPeople - $currentBookingsCount;
+
+        if ($currentBookingsCount >= $maxPeople) {
+            return redirect()->route('bookings.index')
+                ->with('warning', 'Vous avez atteint le nombre maximum d\'inscriptions pour ce créneau horaire.');
+        }
 
         Booking::create([
             'user_id' => Auth::user()->id,
             'schedule_id' => $request->schedule_id
         ]);
 
-        return redirect()->route('bookings.index')->with('success', 'Réservation ajoutée avec succès.');
+        return redirect()->route('bookings.index')
+            ->with('success', 'Vous vous êtes bien inscrit à l\'entraînement ' . strtoupper($schedule->training->name) .
+                ' du ' . date_format(date_create($schedule->start_date), 'd/m/y') . ' à ' .
+                str_replace('h00', 'h', date_format(date_create($schedule->start_date), 'H\hi')) . ' de ' .
+                str_replace('h00', 'h', date_format(date_create($schedule->end_date), 'H\hi')) .
+                '. Il reste ' . $remainingPlaces . ' places pour ce créneau horaire.');
     }
 
     /**
@@ -82,9 +108,6 @@ class BookingController extends Controller
             'schedule_id' => 'required|exists:schedules,id'
         ]);
 
-        // Vérifier si l'utilisateur authentifié est le propriétaire de la réservation
-        // TODO
-        // Auth::check() && $user->role !== 'admin'
         if (Auth::id() !== $booking->user_id || User::find($booking->user_id)->role !== 'admin') {
             return redirect()->route('bookings.index')
                 ->with('danger', 'Vous n\'êtes pas autorisé à modifier cette réservation.');
@@ -104,7 +127,10 @@ class BookingController extends Controller
      */
     public function destroy(Booking $booking): RedirectResponse
     {
-        $booking->delete();
-        return redirect()->route('bookings.index')->with('success', 'Réservation supprimée avec succès.');
+        if (Auth::user()->role === 'admin' || Auth::id() === $booking->user_id) $booking->delete();
+        else return redirect()->route('bookings.index')
+            ->with('danger', 'Vous ne pouvez pas supprimer cette inscription.');
+
+        return redirect()->route('bookings.index')->with('success', 'Inscription supprimée avec succès.');
     }
 }
